Test GetTransaction, UpdateUserRedeem, Earncoin

+++ models/redeem.go
@@ -14,7 +14,7 @@ type TransactionRedeem struct {
 	BeforeAmount float64            `json:"before_amount"bson:"before_amount"`
 	AfterAmount  float64            `json:"after_amount"bson:"after_amount"`
 	SlipNumber   int                `json:"slip_number"bson:"slip_number"`
-	Coin         int                `json:"coin"bson:"coin"`
+	Coin         int64              `json:"coin"bson:"coin"`
 	CreatedAt    time.Time          `json:"created_at"bson:"created_at"`
 	IsRedeem     bool               `json:"is_redeem"bson:"is_redeem"`
 }
