+++ models/setting.go
@@ -25,8 +25,8 @@ type Condition struct {
 	ID     primitive.ObjectID `json:"id"bson:"_id,omitempty"`
 	Prefix string             `json:"prefix"bson:"prefix"`
 	Detail []struct {
-		SlipNumber int `json:"slip_number"bson:"slip_number"`
-		RedeemCoin int `json:"redeem_coin"bson:"redeem_coin"`
+		SlipNumber int   `json:"slip_number"bson:"slip_number"`
+		RedeemCoin int64 `json:"redeem_coin"bson:"redeem_coin"`
 	} `json:"detail"bson:"detail"`
 	MinTopUp   float64   `json:"min_top_up"bson:"min_top_up"`
 	MaxBalance float64   `json:"max_balance"bson:"max_balance"`
